more staff

+++ task-two-object-oriented-c/include/Arayuz.h
@@ -11,16 +11,16 @@
 
 struct Arayuz {
     void* (*cleanScreen)(struct Arayuz*);
-    void* (*writeStatus)(struct Arayuz*, int luckyNumber, int round, float house, struct Kisi* topRich);
-    void* (*writeFinish)(struct Arayuz*,int round, float house);
+    void* (*writeStatus)(struct Arayuz*, int luckyNumber, int round, double house, struct Kisi* topRich);
+    void* (*writeFinish)(struct Arayuz*,int round, double house);
 };
 
 typedef struct Arayuz *Arayuz;
 
 Arayuz new_Arayuz();
 
-void* writeStatus(const Arayuz arayuz, int luckyNumber, int round, float house, Kisi topRich);
-void* writeFinish(const Arayuz arayuz, int round, float house);
+void* writeStatus(const Arayuz arayuz, int luckyNumber, int round, double house, Kisi topRich);
+void* writeFinish(const Arayuz arayuz, int round, double house);
 void* cleanScreen(const Arayuz arayuz);
 
 #endif //PDP_TASK_TWO_ARAYUZ_H
