+++ task-two-object-oriented-c/include/Kisi.h
@@ -10,23 +10,23 @@
 
 struct Kisi{
     char* name;
-    float totalMoney;
-    float spendMoneyEachRound;
+    double totalMoney;
+    double spendMoneyEachRound;
     int luckyNumber;
 
     char* (*getName)(struct Kisi*);
-    float* (*getTotalMoney)(struct Kisi*);
-    float* (*getSpendMoneyEachRound)(struct Kisi*);
+    double* (*getTotalMoney)(struct Kisi*);
+    double* (*getSpendMoneyEachRound)(struct Kisi*);
     int* (*getLuckyNumber)(struct Kisi*);
 };
 
 typedef struct Kisi *Kisi;
 
-Kisi new_Kisi(char* name, float totalMoney, float spendMoneyEachRound, int luckyNumber);
+Kisi new_Kisi(char* name, double totalMoney, double spendMoneyEachRound, int luckyNumber);
 
 char* getName(const Kisi kisi);
-float* getTotalMoney(const Kisi kisi);
-float* getSpendMoneyEachRound(const Kisi kisi);
+double* getTotalMoney(const Kisi kisi);
+double* getSpendMoneyEachRound(const Kisi kisi);
 int* getLuckyNumber(const Kisi kisi);
 
 #endif //PDP_TASK_TWO_KISI_H
